modify the structure of the project to use express (prepare template system)
