+++ webSite/public/scripts/auth.js
@@ -1,0 +1,5 @@
+
+
+$(document).ready(function () {
+
+});
