+++ webSite/public/style.css
@@ -1,0 +1,326 @@
+@import url('https://fonts.googleapis.com/css?family=Work+Sans');
+@import url('https://fonts.googleapis.com/css?family=Shrikhand');
+
+body{
+	margin: 0;
+	padding: 0;
+	background-color:#CCCCCC;
+	overflow: hidden;
+}
+
+/* set the header */
+header{
+	width: 100%;
+	height: 4.375em;
+}
+
+.wrapper{
+	font-family: 'Work Sans', sans-serif;
+	margin: 0 auto;
+}
+
+/* banner background and text properties */
+#banner{
+	width: 100%;
+	background-image:url('img/bg_banner.jpg');
+	font-family: 'Shrikhand', cursive;
+	color: white;
+}
+#bannerText{
+	position:absolute;
+	top:0.2em;
+	left:2em;
+	font-size: 3.125em;
+}
+
+/*resize text if screen too small*/
+@media (max-width: 500px) {
+	#bannerText{font-size: 2em;
+		position:absolute;
+		top:0.4em;
+		left:3em;}
+		#contact{
+			display: none;
+		}
+		.footer{
+			height: 2em;
+		}
+	}
+
+	/* position of banner */
+	.pinpintBanner{
+		width: 100%;
+		height: 5em;
+		vertical-align: top;
+	}
+
+	/*set navbar postion */
+	.navbar{
+		width: 100%;
+		height: auto;
+		padding-bottom: 0.625em;
+		position: fixed;
+	}
+
+	.navbar-nav {
+		font-family: 'Work Sans', sans-serif;
+		font-size: 1.2em;
+		font-weight:bold;
+	}
+
+	.navbar a:hover {
+		background: #696969;
+	}
+
+	/*set logo size */
+	.brand{
+		position: absolute;
+		left: 0.625em;
+		top: 0.625em;
+		z-index: 1;
+		width: 5em;
+		height: 5em;
+	}
+
+	/* login and register links */
+	#register{
+		position: absolute;
+		right: 0.625em;
+		top: 0.625em;
+		width: auto;
+		height: auto;
+		list-style-type:none;
+		display:inline;
+	}
+	.registertxt, .logintxt{
+		margin: 0 5% 0 5%;
+		font-size: 1.2em;
+		color: white;
+		text-align: right;
+		text-decoration: none;
+		font-family: 'Work Sans', sans-serif;
+		font-weight:bold;
+	}
+	.registertxt:hover,.logintxt:hover{
+		background: #696969;
+		cursor: pointer;
+		text-decoration: none;
+	}
+
+	/* main pages css */
+	main{
+		width: 100%;
+		height: auto;
+		float: left;
+		margin: 0;
+		padding: 0;
+	}
+
+	/* set width and height of main content */
+	#container{
+		width: 100%;
+		height: 70%;
+		position: absolute;
+	}
+
+	/* google maps display */
+	#map {
+		height: 100%;
+		width: 100%;
+		position: absolute;
+		z-index: initial;
+	}
+
+	/*search features over google map */
+	.searchfunction{
+		position: absolute;
+		z-index: 1;
+		display: inline;
+		margin: 10px;
+	}
+
+	.HEADINGMAP{
+		text-align: center;
+		margin-top:15%;
+		color: #000;
+	}
+	#overlay{
+		width: 75%;
+		height: 37.5em;
+		background:rgba(107, 107, 107, 1);
+		position: absolute;
+		z-index: -1;
+		left: 30em;
+		top: 6.25em;
+	}
+
+	/* style for about us page */
+	#aboutbox{
+		width:auto;
+		height:auto;
+		/*background-color:#0000FF;*/
+		float:left;
+		margin: 2em;
+		text-align: justify;
+	}
+
+	/* contact us form styles */
+	.contactus{
+		width: 70%;
+		padding: 1em;
+	}
+
+	/* css for the footer and footer componenets */
+	footer{
+		width: 100%;
+		height: 4em;
+		background-color: #1f1f1f;
+		display: block;
+		bottom:0;
+		left:0;
+		color: white;
+		position: fixed;
+	}
+
+	.sm{
+		margin: 0.25em;
+		float: left;
+	}
+
+#contact{
+	top: 20%;
+	text-align: center;
+	position: absolute;
+	width: 100%;
+}
+
+
+
+	.facebook,.twitter{
+		width: 10%;
+		padding: .2em;
+ height: auto;
+		text-decoration: none;
+	}
+	.facebook:hover{
+		opacity: 0.5;
+		cursor: pointer;
+	}
+	.twitter:hover{
+		opacity: 0.5;
+		cursor: pointer;
+	}
+	.copyright{
+		font-size: .8em;
+		padding-right: 0.313em;
+		display: inline;
+		position: absolute;
+		right: 0;
+		top: 0;
+	}
+
+	/* Set a style for all buttons */
+	button {
+		background-color: #4CAF50;
+		color: white;
+		padding: 0.875em 1.25em;
+		margin: 0.5em 0;
+		border: none;
+		cursor: pointer;
+		width: 100%;
+	}
+
+	button:hover {
+		opacity: 0.8;
+	}
+
+	/* Set the loading icon when waiting for map */
+	.location-loading {
+		display: none;
+		position: relative;
+		top: 50%;
+	}
+
+	.ball {
+		background-color: rgba(0, 0, 0, 0);
+		border: 0.313em solid rgba(0, 183, 229, 0.9);
+		opacity: .9;
+		border-top: 0.313em solid rgba(0, 0, 0, 0);
+		border-left: 0.313em solid rgba(0, 0, 0, 0);
+		border-radius: 3.125em;
+		box-shadow: 0 0 2.188em #2187e7;
+		width: 3.125em;
+		height: 3.125em;
+		margin: 0 auto;
+		-moz-animation: spin .5s infinite linear;
+		-webkit-animation: spin .5s infinite linear;
+	}
+
+	.ball1 {
+		background-color: rgba(0, 0, 0, 0);
+		border: 0.313em solid rgba(0, 183, 229, 0.9);
+		opacity: .9;
+		border-top: 0.313em solid rgba(0, 0, 0, 0);
+		border-left: 0.313em solid rgba(0, 0, 0, 0);
+		border-radius: 3.125em;
+		box-shadow: 0 0 0.938em #2187e7;
+		width: 1.875em;
+		height: 1.875em;
+		margin: 0 auto;
+		position: relative;
+		top: -2.5em;
+		-moz-animation: spinoff .5s infinite linear;
+		-webkit-animation: spinoff .5s infinite linear;
+	}
+
+	@-moz-keyframes spin {
+		0% {
+			-moz-transform: rotate(0deg);
+		}
+
+		100% {
+			-moz-transform: rotate(360deg);
+		}
+	}
+
+	@-moz-keyframes spinoff {
+		0% {
+			-moz-transform: rotate(0deg);
+		}
+
+		100% {
+			-moz-transform: rotate(-360deg);
+		}
+	}
+
+	@-webkit-keyframes spin {
+		0% {
+			-webkit-transform: rotate(0deg);
+		}
+
+		100% {
+			-webkit-transform: rotate(360deg);
+		}
+	}
+
+	@-webkit-keyframes spinoff {
+		0% {
+			-webkit-transform: rotate(0deg);
+		}
+
+		100% {
+			-webkit-transform: rotate(-360deg);
+		}
+	}
+
+	.map-error {
+		top: -50%;
+		width: 5%;
+		position: relative;
+		background-color: red;
+	}
+
+	/* header and footer colour of modal */
+	.modal-header, .modal-footer{
+		background-color: #F5F5F5;
+	}
